+++ CHANGELOG
@@ -1,4 +1,7 @@
 0.83.25
+  - Menu options "Force scaler" and "Print text screen"
+    can now be assigned to keyboard shortcuts from the
+    Mapper Editor. (Wengier)
   - Allowed to boot from Toshiba DOS image file for the
     Toshiba J-3100 emulation. (nanshiki)
   - Default S3 machine types now support XGA/SXGA VText
